Move parser test to their own file
